% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_Matrix_Raking.R
\name{compute_D_raking}
\alias{compute_D_raking}
\title{Compute design-based covariance matrix D from a survey design object}
\usage{
compute_D_raking(sp_des, Xp, lonely.psu = "adjust")
}
\arguments{
\item{sp_des}{A survey design object of class "survey.design2" or
"svyrep.design".}

\item{Xp}{Numeric matrix of dimension n_p x p. Each row is h_i.}
}
\value{
A p x p covariance matrix D.
}
\description{
Internal helper to compute
  D = Var_p( sum_{i in sp} d_i * h_i )
when the probability sample is already supplied as a survey design object
created by survey::svydesign() or survey::svrepdesign().
}
\keyword{internal}
