% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPWM_Multi_Estimate.R
\name{multi_estimate}
\alias{multi_estimate}
\title{Estimate step for multi-reference raking (mean + variance)}
\usage{
multi_estimate(Y, Z, w, X, D, S_beta)
}
\arguments{
\item{Y}{Outcome vector (length n_sc_used).}

\item{Z}{Domain/indicator vector (length n_sc_used). Use rep(1, n) if no domain.}

\item{w}{Pseudo-weights from build (length n_sc_used).}

\item{X}{Convenience-sample design matrix (n_sc_used x p), i.e., Xc after NA filtering.}

\item{D}{Block design-based variance matrix from multi reference samples (p x p).}

\item{S_beta}{Score/Information matrix from build: t(w * X) %*% X (p x p).}
}
\value{
list(mean, variance)
}
\description{
Estimate step for multi-reference raking (mean + variance)
}
