% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi_Precalibrate.R
\name{PRECALI}
\alias{PRECALI}
\title{Two-sample pre-calibration function}
\usage{
PRECALI(wts, refs, dup_vars)
}
\arguments{
\item{wts}{A character vector of length 2, giving the names of the weight
columns in the two reference datasets (\code{c(wt1, wt2)}).}

\item{refs}{A list of two data frames. \code{refs[[1]]} is the first reference
sample, and \code{refs[[2]]} is the second reference sample to be calibrated.}

\item{dup_vars}{A character vector of variable names used for calibration.
If empty, calibration is based only on the total weight.}
}
\value{
A numeric vector of calibrated weights for \code{refs[[2]]}.
}
\description{
Performs pre-calibration between two probability samples by aligning
the marginal distributions of selected variables (\code{dup_vars})
in the second reference sample to those of the first one.
}
\details{
The calibration is performed using functions from the \pkg{survey} package:
\code{\link[survey]{svydesign}}, \code{\link[survey]{svytotal}}, and
\code{\link[survey]{calibrate}}. The resulting weights align the margins of
the second reference sample with those of the first reference sample.
}
\examples{
\dontrun{
sp1 <- data.frame(x = rnorm(100), wt1 = runif(100, 0.5, 1.5))
sp2 <- data.frame(x = rnorm(120), wt2 = runif(120, 0.5, 1.5))
PRECALI(wts = c("wt1", "wt2"), refs = list(sp1, sp2), dup_vars = "x")
}

}
\seealso{
[survey::svydesign], [survey::svytotal], [survey::calibrate]
}
