% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Estimate.R
\name{print.IPWM_estimate}
\alias{print.IPWM_estimate}
\title{Print method for IPWM_estimate objects}
\usage{
\method{print}{IPWM_estimate}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"IPWM_estimate"}.}

\item{...}{Additional arguments (not used).}
}
\description{
Displays the pseudo-weighted (adjusted) mean estimate and its uncertainty.
}
